Merge 0.2.4 changes

+++ menus.py
@@ -248,13 +248,17 @@ class CompSetupMenu():
         cList = open("save/compilers.txt", "r")
         cOptions = cList.read().split('\n')
         cOptions.pop(len(cOptions)-1)
+        cList.close()
         self.selComp = "GoldSRC"
-        self.gameSel = ttk.Combobox(master, values=cOptions)
+        self.comps = GamesHandler(cOptions)
+        self.gameSel = ttk.Combobox(master, values=self.comps.gNames)
         self.gameSel.current(0)
         self.gameSel.bind("<<ComboboxSelected>>", self.chComp)
         self.setupLabel = Label(master, text="Compiler Setup", background=thme["bg"], foreground=thme["txt"])
         self.nameLabel = Label(self.top, text="Name: ", background=thme["bg"], foreground=thme["txt"])
-        self.pathLabel = Label(self.top, text="Custom path: ", background=thme["bg"], foreground=thme["txt"])
+        self.pathLabel = Label(self.top, text="Path: ", background=thme["bg"], foreground=thme["txt"])
+        self.typeLabel = Label(self.top, text="Engine Type: ")
+        self.cPathLabel = Label(self.top, text="Custom path: ", background=thme["bg"], foreground=thme["txt"])
         self.name = StringVar()
         self.name.set(cOptions[0])
         self.nameEntry = Entry(self.top, textvariable=self.name, width=50)
@@ -265,6 +269,9 @@ class CompSetupMenu():
         self.csPathButton = Button(self.top, text="Save Path", command=self.savePath)
         if not startHidden:
             self.show()
+
+        self.addGame = Button(self.top, text="Add New Game", command=self.addNComp)
+        self.saveGame = Button(self.top, text="Save Game", command=self.saveComp)
         
         # Applying theme
         self.applyTheme(master)
@@ -302,6 +309,45 @@ class CompSetupMenu():
                     w.configure(fg=self.thme["txt"])
                 except:
                     pass
+    
+    def addNComp(self):
+        pass
+        """self.gameSel.set("")
+        self.name.set("")
+        self.typeSel.current(0)
+        self.hrBool.set(False)
+        self.ucBool.set(False)
+        self.fbBool.set(False)
+        self.newGame = True"""
+    
+    def saveComp(self):
+        pass
+        """if self.newGame:
+            # Newgrounds Reference!?!?!
+            self.nG = self.name.get()
+            if not self.nG.lower() == "goldsrc" or not self.nG.lower() == "svengine":
+                oList = open("save/games.txt", "w")
+                self.gOptions.append(f"{self.nG}~")
+                nList = '\n'.join(self.gOptions)
+                nList = nList + '\n'
+                oList.write(nList)
+                oList.close()
+                uJS = {
+                    self.nG: {
+                        "type": self.typeSel.get(), 
+                        "capabilities": {
+                            "fullbright": self.fbBool.get(),
+                            "1024px": self.hrBool.get(),
+                            "unlockedChrome": self.ucBool.get()
+                        }
+                    }
+                }
+                js = open(f"save/user/comp{self.nG}.json", "w")
+                js.write(json.dumps(uJS, sort_keys=True, indent=5))
+                js.close()
+                self.games = GamesHandler(self.gOptions)
+                self.gameSel["values"] = self.games.gNames
+                self.updFunc(self.games)"""
     
     def inputHandler(self, e=False):
         self.csPath.set(self.csPathEntry.get())
@@ -341,7 +387,7 @@ class CompSetupMenu():
             self.hiddenEdit = False
             self.nameLabel.grid(column=1, row=4, sticky=(W))
             self.nameEntry.grid(column=2, row=4, sticky=(W))
-            self.pathLabel.grid(column=1, row=5, sticky="w")
+            self.cPathLabel.grid(column=1, row=5, sticky="w")
             self.csPathEntry.grid(column=2, row=5, sticky="w")
             self.csPathButton.grid(column=3,row=5,sticky="w", padx=(5,0))
         # If editing options were available and the compiler has editing disabled
@@ -349,7 +395,7 @@ class CompSetupMenu():
             self.hiddenEdit = True
             self.nameLabel.grid_remove()
             self.nameEntry.grid_remove()
-            self.pathLabel.grid(column=1, row=4, sticky="w")
+            self.cPathLabel.grid(column=1, row=4, sticky="w")
             self.csPathEntry.grid(column=2, row=4, sticky="w")
             self.csPathButton.grid(column=3,row=4,sticky="w", padx=(5,0))
 
@@ -366,13 +412,13 @@ class CompSetupMenu():
             self.top.grid(column=1, row=4, sticky="nsew")
             self.nameLabel.grid(column=1, row=0, sticky=(W))
             self.nameEntry.grid(column=2, row=0, sticky=(W))
-            self.pathLabel.grid(column=1, row=1, sticky="w")
+            self.cPathLabel.grid(column=1, row=1, sticky="w")
             self.csPathEntry.grid(column=2, row=1, sticky="w", padx=(15,0))
             self.csPathButton.grid(column=3,row=1,sticky="w", padx=(5,0))
         else:
             self.hiddenEdit = True
             self.top.grid(column=1, row=4, sticky="nsew")
-            self.pathLabel.grid(column=1, row=0, sticky="w")
+            self.cPathLabel.grid(column=1, row=0, sticky="w")
             self.csPathEntry.grid(column=2, row=0, sticky="w", padx=(15,0))
             self.csPathButton.grid(column=3,row=0,sticky="w", padx=(5,0))
 
@@ -452,7 +498,7 @@ class DecompMenu():
         self.logVal = BooleanVar(self.advOpt, value=False)
         self.logChk = Checkbutton(self.advOpt, text="Write log to file", variable=self.logVal, command=self.setLog)
         self.mVal = BooleanVar(self.advOpt, value=self.presetDat["-m"])
-        self.mChk = Checkbutton(self.advOpt, text="GoldSRC compatability", variable=self.mVal)
+        self.mChk = Checkbutton(self.advOpt, text="GoldSRC compatibility", variable=self.mVal)
         self.uVal = BooleanVar(self.advOpt, value=self.presetDat["-u"])
         self.uChk = Checkbutton(self.advOpt, text="Fix UV shifts", variable=self.uVal)
         self.vVal = BooleanVar(self.advOpt, value=self.presetDat["-V"])
@@ -633,6 +679,7 @@ class DecompMenu():
         output = self.out.get()
         gotArgs = False
         cmdArgs = self.getArgs()
+        error = False
         if not cmdArgs == "" or not cmdArgs == " ":
             gotArgs = True
         if output == "" or output == None:
@@ -654,6 +701,25 @@ class DecompMenu():
         # So instead I have to use wine for Mac systems
         """elif sys.platform == 'darwin':
             tOutput = subprocess.getoutput(f'wine third_party/mdldec_win32.exe \"{mdl}\"')"""
+        # Checking for errors (especially the 'unknown Studio MDL format')
+        if tOutput.find("unknown Studio MDL format version 6") != -1:
+            # Telling the file moving part of the function that we are decompiling a v6 MDL file.
+            error = True
+            if sys.platform == 'linux':
+                shutil.copy(mdl, './')
+                tOutput = subprocess.getoutput(f'wine \"{os.getcwd()}/third_party/mdl6dec.exe\" \"{os.path.basename(mdl)}\" -p \"MDL6job\"')
+                os.remove(f"{os.path.basename(mdl)}")
+                # Moving the decompiler output to the output folder!
+                if not os.path.exists(output):
+                    os.mkdir(output)
+                for f in os.listdir('MDL6job'):
+                    print(f)
+                    shutil.copy(f"MDL6job/{f}", os.path.join(output, f))
+                shutil.rmtree('MDL6job')
+            else:
+                tOutput = subprocess.getoutput(f'\"{os.getcwd()}/third_party/mdl6dec.exe\" \"{mdl}\" -p \"{output}\"')
+        elif tOutput.find("ERROR:") != -1:
+            error = True
         print(tOutput)
         self.console.setOutput(tOutput)
         if self.logVal.get():
@@ -663,29 +729,31 @@ class DecompMenu():
             log.write(tOutput)
             log.close()
         # Moving files to output directory (this is a workaround to a bug with Xash3D's model decompiler)
-        filesToMove = []
-        mdlFolder = os.path.dirname(mdl)
-        anims = os.path.join(mdlFolder, 'anims/')
-        texFolder = os.path.join(mdlFolder, 'textures/')
-        for f in os.listdir(mdlFolder):
-            print(f)
-            if f.endswith("smd") or f.endswith("qc"):
-                shutil.copy(f"{mdlFolder}/{f}", os.path.join(output, f))
-                os.remove(f"{mdlFolder}/{f}")
-            elif f.endswith("bmp") and not self.tVal.get():
-                shutil.copy(f"{mdlFolder}/{f}", os.path.join(output, f))
-                os.remove(f"{mdlFolder}/{f}")
-        shutil.copytree(anims, os.path.join(output, 'anims/'))
-        if self.tVal.get():
-            shutil.copytree(texFolder, os.path.join(output, 'textures/'))
+        if not error:
+            if not os.path.exists(output):
+                os.mkdir(output)
+            mdlFolder = os.path.dirname(mdl)
+            anims = os.path.join(mdlFolder, 'anims/')
+            texFolder = os.path.join(mdlFolder, 'textures/')
+            for f in os.listdir(mdlFolder):
+                print(f)
+                if f.endswith("smd") or f.endswith("qc"):
+                    shutil.copy(f"{mdlFolder}/{f}", os.path.join(output, f))
+                    os.remove(f"{mdlFolder}/{f}")
+                elif f.endswith("bmp") and not self.tVal.get():
+                    shutil.copy(f"{mdlFolder}/{f}", os.path.join(output, f))
+                    os.remove(f"{mdlFolder}/{f}")
+            shutil.copytree(anims, os.path.join(output, 'anims/'))
+            if self.tVal.get():
+                shutil.copytree(texFolder, os.path.join(output, 'textures/'))
+                try:
+                    shutil.rmtree(texFolder)
+                except:
+                    pass
             try:
-                shutil.rmtree(texFolder)
+                shutil.rmtree(anims)
             except:
                 pass
-        try:
-            shutil.rmtree(anims)
-        except:
-            pass
 
 class CompMenu():
     def __init__(self, template, master, startHidden:bool=False):
@@ -1124,7 +1192,7 @@ class CompMenu():
                 if not gameDat["capabilities"]["1024px"] and handler.check1024px():
                     warnings.append("WARNING: The selected game does not support textures higher than 512x512, please downscale the offending textures!")
             else:
-                if handler.check1024px() and gameDat["capabilities"]["unlockedChrome"]:
+                if handler.check1024px() and gameDat["capabilities"]["1024px"]:
                     warnings.append("WARNING: The selected compiler does not support textures higher than 512x512, please downscale the offending textures!")
                 elif handler.check1024px():
                     warnings.append("WARNING: The selected compiler and game does not support textures higher than 512x512, please downscale the offending textures!")
@@ -1264,7 +1332,7 @@ class CompMenu():
         cOpts = self.getCompilerOptions()
         # Checking if the QC file supplied uses relative pathing for $cd and $cdtexture as the compiler cannot find the files otherwise
         qcRelChk = QCHandler(mdl)
-        qcRelChk.crowbarFormatCheck()
+        qcRelChk.relPathCheck()
         if qcRelChk.cbarFrmt:
             mdl = qcRelChk.newQCPath
 
@@ -1354,7 +1422,7 @@ class AboutMenu():
         self.ver = vnum.read().replace("(OS)", sys.platform)
         self.setupLabel = Label(master, text=f"Snark {self.ver} by:", background=thme["bg"], foreground=thme["txt"])
         credits = ["PostScript", "\nusing:", "MDLDec by Flying With Gauss", "get_image_size by Paulo Scardine", "TkTooltip by DaedalicEntertainment",
-            "JSONC by John Carter"
+            "JSONC by John Carter", "MDL6Dec by GeckoN"
         ]
         self.nameLabel = Label(master, text="\n".join(credits), background=thme["bg"], fg=thme["txt"])
         # Tooltips
